libusig_shim: change to a shared object (.so)
Before this commit, libusig_shim, which is an untrusted part of codes
written in C for Intel SGX, was built as a static library (.a) and is
linked with `usig/sgx/usig-enclable.go` ("go-wrapper").

This commit makes changes to build the library as a shared object (.so),
as the first step to separate the C codes from the Go code to be able to
build this project in the Go's standard way (i.e. without a Makefile).

Signed-off-by: Yuta Namiki <y-namiki@ak.jp.nec.com>